move preferences atom to separate directory
