+++ core-ui/src/state/preferences/isPreferencesModalOpenAtom.ts
